+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 RELEASE NOTES
 ----------------------------------------------------------------------
 
+## Version 1.0.8, ??-08-2016
+* DoubleClick protocol v.94: AMP support.
+* Fixed `DoubleClickCrypto`'s handling of the initialization vector to
+  be compliant with DoubleClick's recommended encoding of timestamp.
+
 ## Version 1.0.7, 27-07-2016
 * Fixed mapping of `Bid.adm` for non-VAST video ads (e.g. MRAID).